# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Entrophy API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/guilds/{guildId}/integrations": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{17,20}$"
            },
            "in": "path",
            "name": "guildId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/webhooks/github/{endpointId}": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "endpointId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/webhooks/generic/{endpointId}": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "endpointId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}